add types for navigation on the same logi file
